[UI 제작] 플레이어 UI 기능구현
테스트까지 완료

+++ Source/RiseOfSun/Public/RSPlayer.h
@@ -78,7 +78,7 @@ public:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Player Stats", meta = (AllowPrivateAccess = "true"))
     float CurrentHp;
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Player Stats")
-    int32 CurrentEXP;
+    float CurrentEXP;
     UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Player Stats")
     int32 MaxEXP;
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Player Stats")
@@ -94,7 +94,7 @@ public:
 
 	// Getter: 현재 EXP
     UFUNCTION(BlueprintPure, Category = "Player Stats")
-    int32 GetCurrentEXP() const { return CurrentEXP; }
+    float GetCurrentEXP() const { return CurrentEXP; }
 
     // Getter: 최대 EXP
     UFUNCTION(BlueprintPure, Category = "Player Stats")
